requirements.txt
wheel
nest_asyncio
tortoise-orm[aiomysql]
discord.py
git+https://github.com/artem30801/discord-py-slash-commandaiohttp
python-dateutil
fuzzywuzzy
colour
psutil
Pillow